+++ prisma/migrations/20260813205545_lead_no_answer/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "Lead" ADD COLUMN     "noAnswer" BOOLEAN NOT NULL DEFAULT false;